backend/requirements.txt
# Django and REST Framework
Django==5.2.7
djangorestframework==3.14.0
djoser==2.2.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9

# Channels and WebSockets
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Scheduler
django-apscheduler==0.6.2
APScheduler==3.10.4

# LangChain and AI
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-chroma>=0.1.2
langchain-core>=0.3.0
google-generativeai>=0.8.0

# ChromaDB
chromadb>=0.4.22

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Image Generation
runware==0.4.26
